requirements.txt
numpy>=1.26
faiss-cpu>=1.7.4
torch>=2.2
torchvision>=0.17
timm>=0.9
Pillow>=10.0
streamlit>=1.32
